requirements.txt
django
psycopg2
pillow
django-rest-framework